refactor: página de estoque e de análise
