+++ src/main/webapp/css/paginas/admin/analise.css
@@ -1,5 +1,3 @@
-@import url(/css/style.css);
-
 #main-container {
     justify-content: space-around;
     align-content: baseline;
