+++ src/main/webapp/css/paginas/admin/estoque.css
@@ -1,5 +1,3 @@
-@import url(../../../../css/style.css);     
-
 .sessao-data {
     display: inline-flex;
