+++ src/main/webapp/js/paginas/admin/estoque.js
@@ -1,0 +1,66 @@
+import { alertarErro } from "../../api/alertErro.js";
+import { consultarEntradasEstoque } from "../../api/estoque.api.js";
+import { formatarDateTime } from "../../script.js";
+import ModalEntradaEstoque from "@componentes/estoque/modalEntradaEstoque.js";
+import { consultarTodosQuadrinhos } from "@api/quadrinho.api.js";
+import { formatarPreco } from "../../script.js";
+
+export async function initPagina() {
+
+    const tbody = document.getElementById("stockTable");
+
+    let quadrinhos;
+    let entradasEstoque;
+
+    try {
+        entradasEstoque = await consultarEntradasEstoque();
+        quadrinhos = await consultarTodosQuadrinhos();
+    } catch (error){
+        alertarErro(error);
+    }
+
+    const modal = new ModalEntradaEstoque();
+
+    let contador = 1;
+    quadrinhos.forEach(quadrinho => {
+        const tr = document.createElement("tr");
+
+        let estoque = quadrinho.quantidadeEstoque;
+
+        if (quadrinho.quantidadeEstoque < 1){
+            estoque = 'Fora de Estoque';
+        }
+
+        tr.innerHTML = `
+            <td>${contador++}</td>
+            <td class="item-name">${quadrinho.titulo}</td>
+            <td class="item-quantity">${estoque}</td>
+            <td class="item-price">${formatarPreco(quadrinho.preco)}</td>
+            <td class="item-precification">${quadrinho.grupoPrecificacao.nome} ${quadrinho.grupoPrecificacao.porcentagem}%</td>
+            <td>
+                <button class="btn btn-warning btn-sm btn-entrada-estoque">Fazer entrada no estoque</button>
+            </td>
+        `;
+
+        const button = tr.querySelector('button');
+        button.onclick = () => {
+            modal.show(quadrinho);
+        };
+        tbody.append(tr);
+    });
+
+    entradasEstoque.forEach(entrada => {
+        document.getElementById('tabela-entrada-estoque')
+            .insertAdjacentHTML('beforeend', `
+            <tr>
+                <td>#</td>
+                <td>${entrada.nomeQuadrinho}</td>
+                <td>${formatarDateTime(entrada.dataEntrada)}</td>
+                <td>${entrada.quantidade}</td>
+                <td>${entrada.valorCusto}</td>
+                <td>${entrada.fornecedor}</td>
+            </tr>
+            `);
+    });
+
+}
