+++ src/main/webapp/paginas/admin/analise.html
@@ -1,0 +1,45 @@
+<title>Análise</title>
+
+<nav id="main-nav"></nav>
+
+<div id="main-container" class="container d-flex flex-row py-5">  
+    
+    <div class="filtro-analise centered d-flex">
+
+        <form class="form-filtros">
+            <label>
+                Data início
+                <input name="dataInicio" type="date">
+            </label>
+    
+            <label>
+                Data final
+                <input name="dataFinal" type="date">
+            </label>
+        </form>
+
+        <div class="botao">
+            <button id="btn-pesquisar">
+                <img src="/img/search.svg">
+                Filtrar por data
+            </button>
+        </div>
+
+    </div>
+
+    <div id="container-graficos">
+        <canvas id="grafico-linha-quant-produtos" class="grafico-linha"></canvas>
+
+        <hr/>
+
+        <canvas id="grafico-linha-valor-produtos" class="grafico-linha"></canvas>
+
+        <hr/>
+
+        <canvas id="grafico-linha-quant-categorias" class="grafico-linha"></canvas>
+
+        <hr/>
+
+        <canvas id="grafico-linha-valor-categorias" class="grafico-linha"></canvas>
+    </div>
+</div>
