--- src/main/webapp/paginas/admin/analise/analise.html
@@ -1,62 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Análise</title>
-    <link rel="stylesheet" href="../../../paginas/admin/analise/css/analise.css">
-</head>
-
-<body>
-
-    <nav id="main-nav"></nav>
-
-    <div id="main-container" class="container d-flex flex-row py-5">  
-        
-        <div class="filtro-analise centered d-flex">
-
-            <form class="form-filtros">
-                <label>
-                    Data início
-                    <input name="dataInicio" type="date">
-                </label>
-        
-                <label>
-                    Data final
-                    <input name="dataFinal" type="date">
-                </label>
-            </form>
-
-            <div class="botao">
-                <button id="btn-pesquisar">
-                    <img src="/img/search.svg">
-                    Filtrar por data
-                </button>
-            </div>
-
-        </div>
-
-        <div id="container-graficos">
-            <canvas id="grafico-linha-quant-produtos" class="grafico-linha"></canvas>
-
-            <hr/>
-
-            <canvas id="grafico-linha-valor-produtos" class="grafico-linha"></canvas>
-
-            <hr/>
-
-            <canvas id="grafico-linha-quant-categorias" class="grafico-linha"></canvas>
-
-            <hr/>
-
-            <canvas id="grafico-linha-valor-categorias" class="grafico-linha"></canvas>
-        </div>
-    </div>
-
-    <script type="module" src="../../../js/loadAdminPage.js"></script>
-    <script src="../../../js/chart.umd.js"></script>
-    <script type="module" src="../../../paginas/admin/analise/js/analise.js"></script>
-</body>
-
-</html>
