+++ src/main/webapp/paginas/admin/estoque.html
@@ -1,0 +1,41 @@
+<title>Estoque</title>
+
+<nav id="main-nav"></nav>
+
+<div class="container py-5">
+    <h2 class="text-center mb-4">Controle de Estoque</h2>
+
+    <table class="table table-bordered table-striped">
+        <thead class="table-dark">
+            <tr>
+                <th>#</th>
+                <th>Nome</th>
+                <th>Quantidade</th>
+                <th>Valor de Venda</th>
+                <th>Grupo de Precificação (Margem de Lucro %)</th>
+                <th>Operações</th>
+            </tr>
+        </thead>
+        <tbody id="stockTable"></tbody>
+    </table>
+
+    <h2 class="text-center">Entradas no Estoque</h2>
+
+    <table class="table table-bordered table-striped">
+        <thead class="table-dark">
+            <tr>
+                <th>#</th>
+                <th>Quadrinho</th>
+                <th>Data entrada</th>
+                <th>Quantidade</th>
+                <th>Valor de Custo</th>
+                <th>Fornecedor</th>
+            </tr>
+        </thead>
+
+        <tbody
+            id="tabela-entrada-estoque">
+        </tbody>
+    </table>
+
+</div>
