--- src/main/webapp/paginas/admin/estoque/estoque.html
@@ -1,53 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Estoque</title>
-    <link rel="stylesheet" href="../../../paginas/admin/estoque/css/estoque.css">
-</head>
-
-<body class="bg-light">
-
-    <nav id="main-nav"></nav>
-
-    <div class="container py-5">
-        <h2 class="text-center mb-4">Controle de Estoque</h2>
-
-        <table class="table table-bordered table-striped">
-            <thead class="table-dark">
-                <tr>
-                    <th>#</th>
-                    <th>Nome</th>
-                    <th>Quantidade</th>
-                    <th>Valor de Venda</th>
-                    <th>Grupo de Precificação (Margem de Lucro %)</th>
-                    <th>Operações</th>
-                </tr>
-            </thead>
-            <tbody id="stockTable"></tbody>
-        </table>
-
-        <h2 class="text-center">Entradas no Estoque</h2>
-
-        <table class="table table-bordered table-striped">
-            <thead class="table-dark">
-                <tr>
-                    <th>#</th>
-                    <th>Quadrinho</th>
-                    <th>Data entrada</th>
-                    <th>Quantidade</th>
-                    <th>Valor de Custo</th>
-                    <th>Fornecedor</th>
-                </tr>
-            </thead>
-            <tbody id="tabela-entrada-estoque"></tbody>
-        </table>
-    </div>
-
-    <script type="module" src="/js/loadAdminPage.js"></script>
-    <script type="module" src="/paginas/admin/estoque/js/estoque.js"></script>
-</body>
-
-</html>
